backend/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements-dev.ini
#
build==1.2.1
    # via pip-tools
click==8.1.7
    # via pip-tools
iniconfig==2.0.0
    # via pytest
packaging==24.0
    # via
    #   build
    #   pytest
pip-tools==7.4.1
    # via -r requirements-dev.ini
pluggy==1.5.0
    # via pytest
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pytest==8.2.1
    # via
    #   -r requirements-dev.ini
    #   pytest-asyncio
pytest-asyncio==0.23.7
    # via -r requirements-dev.ini
ruff==0.4.5
    # via -r requirements-dev.ini
wheel==0.43.0
    # via pip-tools

PyPDF2==3.0.0  
openpyxl==3.0.10 
google-generativeai==0.8.3
bcrypt==4.0.1
beanie==1.23.0
motor==3.3.2

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
